Moved all configs UI to a separate folder
